variation-commons is now split into three submodules, the poms and imports have been changed to reflect this
