+++ eva-lib/pom.xml
@@ -18,7 +18,7 @@
     <dependencies>
         <dependency>
             <groupId>uk.ac.ebi.eva</groupId>
-            <artifactId>variation-commons</artifactId>
+            <artifactId>core</artifactId>
         </dependency>
         <dependency>
             <groupId>org.postgresql</groupId>
