+++ pom.xml
@@ -47,8 +47,18 @@
             <!-- General dependencies -->
             <dependency>
                 <groupId>uk.ac.ebi.eva</groupId>
-                <artifactId>variation-commons</artifactId>
-                <version>0.4-SNAPSHOT</version>
+                <artifactId>core</artifactId>
+                <version>0.5-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>uk.ac.ebi.eva</groupId>
+                <artifactId>jpa</artifactId>
+                <version>0.5-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>uk.ac.ebi.eva</groupId>
+                <artifactId>mongodb</artifactId>
+                <version>0.5-SNAPSHOT</version>
             </dependency>
             <dependency>
                 <groupId>org.postgresql</groupId>
